+++ sphinx-docs/source/changelog.rst
@@ -8,6 +8,7 @@ v2.2 (TBD)
 
 - Can read molecules from GROAMCS topology
 - Can compute one-hot encoding for different bead type interactions in CG systems
+- Can generate mapped exclusion lists in CG systems
 
 *Enhancements*